backend/src/models/predictors/chronos_predictor.py
"""
Chronos-Bolt time-series predictors.
"""

import logging
import pandas as pd
import numpy as np
from datetime import datetime, timedelta
from typing import Dict, List, Any

from models.base_predictor import TimeSeriesPredictor


logger = logging.getLogger(__name__)


class ChronosBoltPredictor(TimeSeriesPredictor):
    """Amazon Chronos-Bolt model via HuggingFace - 250x faster than original!"""
    
    def __init__(self, model_size: str = "small"):
        super().__init__(f"Chronos-Bolt-{model_size}")
        self.model_size = model_size
        self.pipeline = None
        self.historical_data = None
        self.predictions = None
        
    def train(self, data: pd.DataFrame) -> None:
        """Load Chronos-Bolt model (pre-trained, no additional training needed)."""
        try:
            from chronos import BaseChronosPipeline
            import torch
            
            logger.info("Loading %s model...", self.name)
            
            # Use new Chronos-Bolt models (much faster!)
            model_name = f"amazon/chronos-bolt-{self.model_size}"
            
            # Optimize for M1 Mac
            device_map = "mps" if torch.backends.mps.is_available() else "cpu"
            torch_dtype = torch.bfloat16 if device_map != "cpu" else torch.float32
            
            logger.info("Using device: %s", device_map)
            
            self.pipeline = BaseChronosPipeline.from_pretrained(
                model_name,
                device_map=device_map,
                torch_dtype=torch_dtype,
            )
            
            # Store historical data
            self.historical_data = data['Close'].values
            self.dates = data.index
            self.is_trained = True
            
        except ImportError:
            raise ImportError("Install chronos-forecasting: pip install chronos-forecasting")
    
    def predict(self, forecast_days: int) -> Dict[str, Any]:
        """Generate Chronos-Bolt predictions."""
        if not self.is_trained:
            raise ValueError("Model must be trained before prediction")
        
        import torch
        
        logger.info("Generating %s-day forecast with %s...", forecast_days, self.name)
        
        # Chronos-Bolt uses direct multi-step forecasting - much faster!
        # Prepare input context tensor
        context = torch.tensor(self.historical_data, dtype=torch.float32)
        
        # Generate forecast (returns quantile forecasts directly)
        with torch.no_grad():
            forecast = self.pipeline.predict(
                context=context,
                prediction_length=forecast_days
            )
        
        # forecast shape: [num_series, num_quantiles, prediction_length]
        # Extract median (50th percentile) and confidence intervals
        forecast_median = forecast[0, len(forecast[0])//2, :].numpy()  # Middle quantile (median)
        forecast_lower = forecast[0, len(forecast[0])//10, :].numpy()  # ~10th percentile
        forecast_upper = forecast[0, -len(forecast[0])//10, :].numpy()  # ~90th percentile
        
        # Generate future dates
        last_date = self.dates[-1]
        forecast_dates = [
            (last_date + timedelta(days=i+1)).strftime('%Y-%m-%d') 
            for i in range(forecast_days)
        ]
        
        self.predictions = {
            'forecast_median': forecast_median,
            'forecast_lower': forecast_lower,
            'forecast_upper': forecast_upper
        }
        
        logger.info("%s prediction complete!", self.name)
        
        return {
            'forecast_dates': forecast_dates,
            'forecast_prices': forecast_median.tolist(),
            'confidence_lower': forecast_lower.tolist(),
            'confidence_upper': forecast_upper.tolist()
        }
    
    def get_metrics(self) -> Dict[str, float]:
        """Calculate meaningful metrics for Chronos-Bolt including prediction accuracy."""
        if not self.predictions or self.historical_data is None:
            return {
                'data_points': len(self.historical_data) if self.historical_data is not None else 0,
                'mae': 0.0,
                'rmse': 0.0, 
                'mape': 0.0
            }
        
        try:
            # For Chronos models, we can't calculate true accuracy metrics since we don't have
            # historical predictions to compare against. Instead, we'll estimate based on 
            # forecast confidence and historical volatility.
            
            data_points = len(self.historical_data)
            
            # Calculate historical volatility as a proxy for expected prediction error
            historical_std = np.std(self.historical_data)
            current_price = self.historical_data[-1]
            
            # Estimate MAE based on historical volatility (typically 60-80% of std dev)
            estimated_mae = historical_std * 0.7
            
            # Estimate MAPE as percentage of current price
            estimated_mape = (estimated_mae / current_price) * 100
            
            # Estimate RMSE (typically 1.2-1.4x MAE)
            estimated_rmse = estimated_mae * 1.3
            
            # If we have confidence intervals, use those to refine estimates
            if 'forecast_lower' in self.predictions and 'forecast_upper' in self.predictions:
                confidence_range = np.mean(np.array(self.predictions['forecast_upper']) - 
                                         np.array(self.predictions['forecast_lower']))
                # Confidence range typically represents ~1.6 standard deviations (80% interval)
                estimated_mae = confidence_range / 3.2  # Convert to MAE estimate
                estimated_mape = (estimated_mae / current_price) * 100
                estimated_rmse = estimated_mae * 1.25
            
            return {
                'mae': float(estimated_mae),
                'rmse': float(estimated_rmse),
                'mape': float(min(estimated_mape, 100.0)),  # Cap at 100%
                'data_points': data_points,
                'model_type': f'chronos-bolt-{self.model_size}',
                'confidence_based': True  # Flag to indicate these are estimates
            }
            
        except Exception as e:
            logger.warning("Error calculating Chronos metrics: %s", e)
            return {
                'data_points': len(self.historical_data) if self.historical_data is not None else 0,
                'mae': 0.0,
                'rmse': 0.0,
                'mape': 0.0
            }
