backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-analysis-platform"
version = "0.1.0"
description = "stock-analysis-platform backend: FastAPI service for financial data, caching, predictions, and LLM analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [	
	# Core ML 
	"pandas",
	"numpy",
	"scikit-learn",
	"joblib",

	# Database
	"sqlalchemy>=2.0",
	"alembic>=1.12",

	# API
	"fastapi",
	"uvicorn[standard]",
	"requests",



	# Configuration
	"pydantic>=2.0",
	"pydantic-settings>=2.0",
	"python-dotenv",

	# Financial Data & Time Series
	"yfinance>=0.2.18",
	"chronos-forecasting",  # Amazon Chronos-Bolt models
	"ta>=0.10.2",           # Technical Analysis indicators
	
	# Optional deep learning dependencies (for advanced models)
	# "torch",              # PyTorch - only needed for Chronos models
	# "einops>=0.6.0",     # Tensor operations - only needed for advanced models
		
	# Text Processing (for ticker resolution & sentiment)
	"beautifulsoup4",
	"requests-html",
	"fuzzywuzzy",
	"python-levenshtein",
]

[project.optional-dependencies]
dev = [
	"pytest",
	"httpx",
	"ruff",
	"black",
	"mypy",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["data*", "features*", "models*", "api*", "services*", "database*", "config*", "llm*"]
exclude = ["tests*", "frontend*", "artifacts*"]
